feat: 🎸 activity rich text optimization
