+++ src/pages/Activity/components/LeaderboardEntry/index.tsx
@@ -5,6 +5,7 @@ import { useCmsTranslations } from 'hooks/cms';
 import { TLeaderboardEntryInfoTranslations } from 'graphqlServer/queries/activity/leaderboardEntry';
 import { LeaderboardEntrySub } from './components/LeaderboardEntrySub';
 import { useMobile } from 'utils/isMobile';
+import { ActivityRichText } from '../common/ActivityRichText';
 
 export type TLeaderboardEntryProps = {
   activity: ILeaderboardEntryActivity;
@@ -32,9 +33,10 @@ export const LeaderboardEntry = ({ activity }: TLeaderboardEntryProps) => {
             <div className="leaderboard-entry-hero-section-title">{t('title')}</div>
 
             {t('description') && (
-              <div
+              <ActivityRichText
                 className="leaderboard-entry-hero-section-description"
-                dangerouslySetInnerHTML={{ __html: t('description') || '' }}></div>
+                innerHTML={t('description') || ''}
+              />
             )}
           </div>
         </div>
